+++ libs/spark/CHANGELOG.md
@@ -9,8 +9,8 @@
 
 ### Features
 
-- **useDropdownContext**
-  - Added return value property, `isOpen: boolean` to know whether the dropdown is open or not.
+- **DropdownAnchor**
+  - When component is **Unstable_Button**, adds default trailing icon and styles (to override, simply set `trailingIcon`).
 - **DropdownMenu**
   - Allow overriding component.
   - Improve offset from Dropdown Anchor when open.
@@ -18,6 +18,8 @@
     - `component`: added, default of `Menu`.
 - **Unstable_SwitchField**
   - Increase label font weight.
+- **useDropdownContext**
+  - Added return value property, `isOpen: boolean` to know whether the dropdown is open or not.
 
 ## [v2.0.0-alpha.3](https://github.com/prenda-school/prenda-spark/compare/v2.0.0-alpha.2...v2.0.0-alpha.3) (2022-07-11)